Add files via upload
It is a use case specification for static diagram, This document is prepared is because to give a clear understanding to developers when developing the application. FRD, UML and specifications document will play a important role to communicate between the business analyst, developers and client. That helps to represent the vision.